backing out due to orange
UltraBlame original commit: c46442f690438b44de0ec653992b0a487a42d0c7

+++ netwerk/cookie/src/nsCookieService.cpp
@@ -54,7 +54,6 @@
 #include "nsIFile.h"
 #include "nsIObserverService.h"
 #include "nsILineInputStream.h"
-#include "nsIEffectiveTLDService.h"
 
 #include "nsCOMArray.h"
 #include "nsArrayEnumerator.h"
@@ -63,6 +62,7 @@
 #include "nsCRT.h"
 #include "prtime.h"
 #include "prprf.h"
+#include "prnetdb.h"
 #include "nsNetUtil.h"
 #include "nsNetCID.h"
 #include "nsAppDirectoryServiceDefs.h"
@@ -422,10 +422,6 @@ nsCookieService::Init()
     return NS_ERROR_OUT_OF_MEMORY;
   }
 
-  nsresult rv;
-  mTLDService = do_GetService(NS_EFFECTIVETLDSERVICE_CONTRACTID, &rv);
-  NS_ENSURE_SUCCESS(rv, rv);
-
   
   nsCOMPtr<nsIPrefBranch2> prefBranch = do_GetService(NS_PREFSERVICE_CONTRACTID);
   if (prefBranch) {
@@ -437,7 +433,7 @@ nsCookieService::Init()
 
   
   
-  rv = InitDB();
+  nsresult rv = InitDB();
   if (NS_FAILED(rv))
     COOKIE_LOGSTRING(PR_LOG_WARNING, ("Init(): InitDB() gave error %x", rv));
 
@@ -615,9 +611,11 @@ nsCookieService::Observe(nsISupports     *aSubject,
 
     if (!nsCRT::strcmp(aData, NS_LITERAL_STRING("shutdown-cleanse").get()) && mDBConn) {
       
-      nsresult rv = mDBConn->ExecuteSimpleSQL(NS_LITERAL_CSTRING("DELETE FROM moz_cookies"));
-      if (NS_FAILED(rv))
-        NS_WARNING("db delete failed");
+      if (mDBConn) {
+        nsresult rv = mDBConn->ExecuteSimpleSQL(NS_LITERAL_CSTRING("DELETE FROM moz_cookies"));
+        if (NS_FAILED(rv))
+          NS_WARNING("db delete failed");
+      }
     }
 
   } else if (!strcmp(aTopic, "profile-do-change")) {
@@ -1130,6 +1128,7 @@ nsCookieService::GetCookieInternal(nsIURI      *aHostURI,
   
   
   hostFromURI.Insert(NS_LITERAL_CSTRING("."), 0);
+  ToLowerCase(hostFromURI);
 
   
   
@@ -1701,12 +1700,73 @@ nsCookieService::ParseAttributes(nsDependentCString &aCookieHeader,
 
 
 
+
+PRBool
+nsCookieService::IsIPAddress(const nsAFlatCString &aHost)
+{
+  PRNetAddr addr;
+  return (PR_StringToNetAddr(aHost.get(), &addr) == PR_SUCCESS);
+}
+
+PRBool
+nsCookieService::IsInDomain(const nsACString &aDomain,
+                            const nsACString &aHost,
+                            PRBool           aIsDomain)
+{
+  
+  
+  
+  if (!aIsDomain) {
+    return aDomain.Equals(aHost);
+  }
+
+  
+  
+
+
+
+
+
+
+
+
+  
+  PRUint32 domainLength = aDomain.Length();
+  PRInt32 lengthDifference = aHost.Length() - domainLength;
+  
+  
+  
+  
+  if (lengthDifference == 0) {
+    return aDomain.Equals(aHost);
+  }
+  
+  if (lengthDifference > 0) {
+    return aDomain.Equals(Substring(aHost, lengthDifference, domainLength));
+  }
+  
+  if (lengthDifference == -1) {
+    return Substring(aDomain, 1, domainLength - 1).Equals(aHost);
+  }
+  
+  return PR_FALSE;
+}
+
 PRBool
 nsCookieService::IsForeign(nsIURI *aHostURI,
                            nsIURI *aFirstURI)
 {
   
   if (!aFirstURI) {
+    return PR_FALSE;
+  }
+
+  
+  
+  
+  PRBool isChrome = PR_FALSE;
+  nsresult rv = aFirstURI->SchemeIs("chrome", &isChrome);
+  if (NS_SUCCEEDED(rv) && isChrome) {
     return PR_FALSE;
   }
 
@@ -1719,53 +1779,46 @@ nsCookieService::IsForeign(nsIURI *aHostURI,
   
   currentHost.Trim(".");
   firstHost.Trim(".");
+  ToLowerCase(currentHost);
+  ToLowerCase(firstHost);
+
+  
+  
 
   
   
   
-  
-  
-  
-  
-  
-  
-  
-  
-  if (firstHost.Equals(currentHost))
-    return PR_FALSE;
-
-  
-  
-  
-  PRBool isChrome = PR_FALSE;
-  nsresult rv = aFirstURI->SchemeIs("chrome", &isChrome);
-  if (NS_SUCCEEDED(rv) && isChrome) {
-    return PR_FALSE;
+  if (IsIPAddress(firstHost)) {
+    return !IsInDomain(firstHost, currentHost, PR_FALSE);
   }
 
   
   
-  nsCAutoString baseDomain;
-  rv = mTLDService->GetBaseDomain(aFirstURI, 0, baseDomain);
-  if (NS_FAILED(rv)) {
+  
+  
+  
+  
+  
+  
+  
+  PRUint32 dotsInFirstHost = firstHost.CountChar('.');
+  if (dotsInFirstHost == currentHost.CountChar('.') &&
+      dotsInFirstHost >= 2) {
     
-    return PR_TRUE;
-  }  
-  baseDomain.Trim(".");
+    PRInt32 dot1 = firstHost.FindChar('.');
+    return !IsInDomain(Substring(firstHost, dot1, firstHost.Length() - dot1), currentHost);
+  }
 
   
   
-  
-  currentHost.Insert(NS_LITERAL_CSTRING("."), 0);
-  baseDomain.Insert(NS_LITERAL_CSTRING("."), 0);
-  return !StringEndsWith(currentHost, baseDomain);
+  return !IsInDomain(NS_LITERAL_CSTRING(".") + firstHost, currentHost);
 }
 
 PRUint32
-nsCookieService::CheckPrefs(nsIURI     *aHostURI,
-                            nsIURI     *aFirstURI,
-                            nsIChannel *aChannel,
-                            const char *aCookieHeader)
+nsCookieService::CheckPrefs(nsIURI         *aHostURI,
+                            nsIURI         *aFirstURI,
+                            nsIChannel     *aChannel,
+                            const char     *aCookieHeader)
 {
   
   
@@ -1842,8 +1895,6 @@ PRBool
 nsCookieService::CheckDomain(nsCookieAttributes &aCookieAttributes,
                              nsIURI             *aHostURI)
 {
-  nsresult rv;
-
   
   nsCAutoString hostFromURI;
   if (NS_FAILED(aHostURI->GetAsciiHost(hostFromURI))) {
@@ -1851,6 +1902,7 @@ nsCookieService::CheckDomain(nsCookieAttributes &aCookieAttributes,
   }
   
   hostFromURI.Trim(".");
+  ToLowerCase(hostFromURI);
 
   
   if (!aCookieAttributes.host.IsEmpty()) {
@@ -1861,27 +1913,29 @@ nsCookieService::CheckDomain(nsCookieAttributes &aCookieAttributes,
     
     
     
-    nsCAutoString baseDomain;
-    rv = mTLDService->GetBaseDomain(aHostURI, 0, baseDomain);
-    baseDomain.Trim(".");
-    if (NS_FAILED(rv)) {
-      
-      
-      
-      
-      
-      if (rv == NS_ERROR_HOST_IS_IP_ADDRESS)
-        return hostFromURI.Equals(aCookieAttributes.host);
+    
+    
+    if (IsIPAddress(aCookieAttributes.host)) {
+      return IsInDomain(aCookieAttributes.host, hostFromURI, PR_FALSE);
+    }
 
+    
+
+
+
+
+
+    PRInt32 dot = aCookieAttributes.host.FindChar('.');
+    if (dot == kNotFound) {
+      
       return PR_FALSE;
     }
 
     
-    
-    
     aCookieAttributes.host.Insert(NS_LITERAL_CSTRING("."), 0);
-    baseDomain.Insert(NS_LITERAL_CSTRING("."), 0);
-    return StringEndsWith(aCookieAttributes.host, baseDomain);
+    if (!IsInDomain(aCookieAttributes.host, hostFromURI)) {
+      return PR_FALSE;
+    }
 
     
 
@@ -1890,18 +1944,18 @@ nsCookieService::CheckDomain(nsCookieAttributes &aCookieAttributes,
 
 
 
-  }
 
   
-  if (hostFromURI.IsEmpty()) {
-    PRBool isFileURI = PR_FALSE;
-    aHostURI->SchemeIs("file", &isFileURI);
-    if (!isFileURI)
-      return PR_FALSE;
+  } else {
+    
+    if (hostFromURI.IsEmpty()) {
+      PRBool isFileURI = PR_FALSE;
+      aHostURI->SchemeIs("file", &isFileURI);
+      if (!isFileURI)
+        return PR_FALSE;
+    }
+    aCookieAttributes.host = hostFromURI;
   }
-
-  
-  aCookieAttributes.host = hostFromURI;
 
   return PR_TRUE;
 }
